+++ src/tracking.rs
@@ -1,6 +1,6 @@
 pub trait SearchStatistics {
     fn hit_a_state(&self, depth: usize);
-    fn hit_unique_state(&self, depth: usize, n_moves: usize);
+    fn hit_unique_state(&self, depth: usize, n_moves: u32);
     fn finish_move(&self, depth: usize);
 }
 
@@ -8,7 +8,7 @@ pub struct EmptySearchStats;
 
 impl SearchStatistics for EmptySearchStats {
     fn hit_a_state(&self, _: usize) {}
-    fn hit_unique_state(&self, _: usize, _: usize) {}
+    fn hit_unique_state(&self, _: usize, _: u32) {}
     fn finish_move(&self, _: usize) {}
 }
